pyproject.toml
[project]
name = "crudfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "cross-web==0.4.1",
    "fastapi==0.128.0",
    "graphql-core==3.2.7",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "idna==3.11",
    "packaging==25.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "strawberry-graphql==0.288.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.40.0",
]
